searchbar, ikke ferdig

+++ app/src/main/java/com/example/in2000_prosjekt/ui/data/MapCoordinatesBuild.kt
@@ -13,7 +13,7 @@ data class Content(
 )
 
 data class Geo(
-    val coordinates : List<Float>?,
+    val coordinates : List<Double>?,
     val type : String?
 )
